Refactor for Tennsy
